% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_.R
\name{draw_}
\alias{draw_}
\alias{draw_barchart}
\alias{draw_heatmap}
\alias{draw_histogram}
\title{Create bar plot with distribution of studies over the region from lat/long info}
\usage{
draw_barchart(df, location_column, axis_txt_lim = 60)

draw_heatmap(idata, selcols, axis_txt_lim = 60)

draw_histogram(idata, hist_col, axis_txt_lim = 60)
}
\arguments{
\item{df}{Input dataframe}

\item{location_column}{Column with location information (preferably country-level or higher)}

\item{idata}{Input dataframe}

\item{selcols}{Numeric vector of column indices}

\item{verbose}{Comments}
}
\value{
Returns a bar plot object showing counts of literature in systematic review for each location

Returns a heatmap object showing number of literature under different categories in user specified \code{selcols}
}
\description{
Feed in a dataset and a plot of the correct type
}
\author{
Sanita Dhaubanjar

Ezgi Tanriver-Ayder and Sanita Dhaubanjar
}
\keyword{SystematicReview}
