% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eviatlas_pilotdata.R
\docType{data}
\name{eviatlas_pilotdata}
\alias{eviatlas_pilotdata}
\title{Eviatlas Sample Data}
\format{
An object of class \code{data.frame} with 1072 rows and 94 columns.
}
\usage{
eviatlas_pilotdata
}
\description{
A sample systematic map dataset used to show features of Eviatlas and
its functions to build a map atlas
}
\keyword{datasets}
