% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenLocationTrend.R
\name{GenLocationTrend}
\alias{GenLocationTrend}
\title{Simple R code to create bar plot with distribution of studies over the region from lat/long info
Created For	  : ES Hackathon 2018}
\usage{
GenLocationTrend(idata, idxlonlat, selcolidx, verbose = FALSE)
}
\arguments{
\item{idata}{Input dataframe}

\item{idxlonlat}{Input data frame with two columns for longitude and latitude in degrees}

\item{selcolidx}{Index for column with location information}

\item{verbose}{Logical variable (default=FALSE) for displaying messages on console or not}
}
\value{
Returns a bar plot object showing number of literature in different locations
}
\description{
Simple R code to create bar plot with distribution of studies over the region from lat/long info
Created For	  : ES Hackathon 2018
}
\examples{

}
\author{
Sanita Dhaubanjar
}
