% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenTimeTrend.R
\name{GenTimeTrend.default}
\alias{GenTimeTrend.default}
\title{Create Histogram from dataset - default method}
\usage{
\method{GenTimeTrend}{default}(idata, hist_col, axis_txt_lim = 60)
}
\description{
Create Histogram from dataset - default method
}
\keyword{internal}
