% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_coord_cols.R
\name{get_latitude_cols}
\alias{get_latitude_cols}
\title{functions to determine if columns contain what appear to be latitude or longitude columns}
\usage{
get_latitude_cols(df)
}
\description{
functions to determine if columns contain what appear to be latitude or longitude columns
}
\keyword{internal}
