% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_link_cols.R
\name{get_link_cols}
\alias{get_link_cols}
\title{get link col
Created For ES Hackathon 2018}
\usage{
get_link_cols(df)
}
\arguments{
\item{df}{input dataframe}
}
\description{
get link col
Created For ES Hackathon 2018
}
\keyword{SystematicReview}
