% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_coord_cols.R
\name{get_longitude_cols}
\alias{get_longitude_cols}
\title{get longitude
Created For ES Hackathon 2018}
\usage{
get_longitude_cols(df)
}
\arguments{
\item{df}{Input dataframe}
}
\description{
get longitude
Created For ES Hackathon 2018
}
