% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_obs.R
\name{get_obs}
\alias{get_obs}
\title{Wrangle data into observations}
\usage{
get_obs(df, var, new_var, sep)
}
\arguments{
\item{df}{A data frame.}

\item{var}{The variable to unnest as a character string.}

\item{new_var}{The name of the new variable as a character string.}

\item{sep}{The character that separates the values.}
}
\description{
This function converts a nested variable, separated by a specified character,
to a long-form dataset, where each row is an observation.
}
