% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pilotdata.R
\docType{data}
\name{pilotdata}
\alias{pilotdata}
\title{Development dataset}
\format{A dataframe.
\describe{
  \item{variables}{We don't need to fully document this dataset as this is a placeholder dataset.}
}}
\usage{
pilotdata
}
\description{
This hasn't been published yet, so we'll need to source a different dataset for publication of the `eviatlas` package.
}
\keyword{datasets}
