% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runShiny.R
\name{runShiny}
\alias{runShiny}
\title{A wrapper function to run Shiny Apps from \code{eviatlas}.}
\usage{
runShiny(app = "eviatlas")
}
\arguments{
\item{app}{eviatlas}
}
\value{
eviatlas shiny app
}
\description{
Running this function will launch the eviatlas shiny
}
