% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_shiny_app.R
\name{save_shiny_app}
\alias{save_shiny_app}
\title{Save the Shiny App to an app.R file.}
\usage{
save_shiny_app(
  ui_function,
  server_function,
  pkgs,
  out_dir = "shiny",
  run_app = interactive()
)
}
\arguments{
\item{ui_function}{the function for the UI.}

\item{server_function}{the function for the server.}

\item{pkgs}{any packages that need to be loaded for the app to work. At
minimum the package containing the shiny app should be included.}

\item{out_dir}{the directory to save the shiny app files.}

\item{run_app}{whether to run the app once the files are}
}
\description{
This function will create three files in the \code{out_dir}: \code{server.R},
\code{ui.R}, and \code{global.R}. The contents of \code{server.R} and
\code{ui.R} will be the source code of the \code{server_function} and
\code{ui_function}, respectively. The \code{global.R} file will only contain
\code{library} calls for \code{shiny} and any other packages specified in
the \code{pkgs} parameter.
}
\details{
If \code{run_app = TRUE} the function will start the Shiny app once the
files are written. This is recommended to ensure all the necessary packages
are loaded for the Shiny app to run.
}
