% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_config.R
\docType{data}
\name{shiny_config}
\alias{shiny_config}
\title{EviAtlas ShinyApp Configuration file}
\format{
An object of class \code{list} of length 1.
}
\usage{
shiny_config
}
\description{
EviAtlas ShinyApp Configuration file
}
\keyword{datasets}
