% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sys_map.R
\name{sys_map}
\alias{sys_map}
\title{Map the studies included in a systematic review.}
\usage{
sys_map(studies_data)
}
\arguments{
\item{studies_data}{Input dataframe}
}
\description{
Created For	  : ES Hackathon 2018
}
\keyword{SystematicReview}
