% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sys_map_shapefile.R
\name{sys_map_shapefile}
\alias{sys_map_shapefile}
\title{generate systematic map atlas for shapefiles}
\usage{
sys_map_shapefile(shp, popups = "")
}
\arguments{
\item{shp}{the shapefile used to generate atlas}

\item{popups}{columns to appear in popups when clicking atlas points}
}
\description{
generate systematic map atlas for shapefiles
}
\keyword{internal}
