\name{vegetated_strips_dataset}
\alias{eviatlas_pilotdata}
\title{Vegetation Strips Systematic Map dataset}
\description{Example systematic map database}
\format{A data.frame containing 1072 observations of 15 variables:
  \describe{
    \item{short_title}{ID for each study}
    \item{title}{Title of published article}
    \item{year}{Year published}
    \item{period}{Five-year window during which this study was published}
    \item{google_scholar_link}{Link to the article on Google Scholar}
    \item{nation}{Country where the study was located}
    \item{study_location}{Study location in detail}
    \item{latitude}{Latitude}
    \item{longitude}{Longitude}
    \item{study_design}{Study design}
    \item{measurement_quarter}{Part of year in which the study occurred}
    \item{farming_system}{Description of the farming system}
    \item{farming_production_system}{Description of the farming production system}
    \item{vegetation}{Vegetation type}
    \item{strip_location}{Location of the vegetation strips investigated in this study}
  }
}
\source{NR Haddaway. C Brown, J Eales, S Eggers, J Josefsson, B Kronvang, NP Randall & J Uusi-Kamppa (2018) The multifunctional roles of vegetated strips around and within agricultural fields. Environmental Evidence 7(1). https://doi.org/10.1186/s13750-018-0126-2
}
