% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eviatlas.R
\name{eviatlas}
\alias{eviatlas}
\title{Eviatlas Shiny App}
\usage{
eviatlas(study_data = NULL, ...)
}
\arguments{
\item{study_data}{a dataframe or tibble containing data to be used in an eviatlas shiny app}

\item{...}{additional variables to feed into shiny app}
}
\description{
Eviatlas Shiny App
}
