% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eviatlas.R
\name{eviatlas}
\alias{eviatlas}
\title{Run a shiny app for visualising systematic map databases}
\usage{
eviatlas(name = "eviatlas_app", title = "eviatlas", data)
}
\arguments{
\item{name}{character giving the name of app to be built}

\item{title}{label to be shown in the app header}

\item{data}{optional data.frame to replace the default dataset}
}
\value{
This function builds an app in the working directory.
}
\description{
eviatlas allows users to create a suite of visualisations from a database of
studies, including Evidence Atlases (interactive geographical maps showing
studies and their details over space), Heat Maps (cross tabulations of
categorical variables that highlight clusters and gaps in the evidence),
descriptive plots that help to visualise the evidence base (e.g. the number
of publications per year), and human-readable databases that are easily
filterable.
}
\examples{
  \dontrun{eviatlas(); shiny::runApp("eviatlas_app")}

}
