% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eviatlas.R
\name{eviatlas}
\alias{eviatlas}
\title{Run a shiny app for visualising systematic map databases}
\usage{
eviatlas(data, name = "eviatlas_app")
}
\arguments{
\item{data}{An optional data.frame to display using \code{eviatlas}}

\item{name}{character giving the name of app to be built}
}
\value{
This function builds an app in the working directory.
}
\description{
eviatlas allows users to create a suite of visualisations from a database of
studies, including Evidence Atlases (interactive geographical maps showing
studies and their details over space), Heat Maps (cross tabulations of
categorical variables that highlight clusters and gaps in the evidence),
descriptive plots that help to visualise the evidence base (e.g. the number
of publications per year), and human-readable databases that are easily
filterable.
}
\examples{
  \dontrun{eviatlas()}

}
