% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenHeatMap.R
\name{GenHeatMap}
\alias{GenHeatMap}
\title{Create heatmap from dataset}
\usage{
GenHeatMap(idata, selcols, verbose = FALSE)
}
\arguments{
\item{idata}{Input data frame}

\item{selcols}{Vector of strings with column names}
}
\value{
Returns a heatmap object showing number of literature under different categories in user specified \code{selcols}
}
\description{
Feed in a dataset and get a heatmap in return
Created For	  : ES Hackathon 2018
}
\examples{
## load the data sets
GenHeatMap(mpg, colnames(mpg[c(1,4)]))

}
\author{
Sanita Dhaubanjar 24 April 2018
}
