% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenLocationTrend.R
\name{GenLocationTrend}
\alias{GenLocationTrend}
\title{Create bar plot with distribution of studies over the region from lat/long info}
\usage{
GenLocationTrend(df, location_column, axis_txt_lim = 60)
}
\arguments{
\item{df}{Input dataframe}

\item{location_column}{Column with location information (preferably country-level or higher)}
}
\value{
Returns a bar plot object showing counts of literature in systematic review for each location
}
\description{
Created For	  : ES Hackathon 2018
}
\author{
Sanita Dhaubanjar
}
\keyword{SystematicReview}
