% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenTimeTrend.R
\name{GenTimeTrend.numeric}
\alias{GenTimeTrend.numeric}
\title{Create Histogram from dataset - numeric column}
\usage{
\method{GenTimeTrend}{numeric}(idata, hist_col, axis_txt_lim = 60)
}
\description{
Create Histogram from dataset - numeric column
}
\keyword{internal}
