% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenTimeTrend.R
\name{GenTimeTrend}
\alias{GenTimeTrend}
\title{Create GenTimeTrend from dataset}
\usage{
GenTimeTrend(idata = pilotdata, year_column = NULL, verbose = FALSE)
}
\arguments{
\item{idata}{Input data frame}

\item{year_column}{List of strings with column names}

\item{df}{A dataframe}

\item{year_column}{a vector of columns to use}
}
\value{
Returns a timetrend object showing number of literature under different categories in user specified \code{year_column}
}
\description{
Created For	  : ES Hackathon 2018
}
\author{
Ezgi Tanriver-Ayder 24 April 2018
}
\keyword{SystematicReview,}
