% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_histogram_viable_cols.R
\name{get_histogram_viable_columns}
\alias{get_histogram_viable_columns}
\title{Function to determine if columns should be considered for use in a histogram. very generous limit of 100.}
\usage{
get_histogram_viable_columns(df)
}
\arguments{
\item{df}{Input dataframe}
}
\description{
Created For ES Hackathon 2018
}
\keyword{SystematicReview}
