% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_link_cols.R
\name{is_link_col}
\alias{is_link_col}
\title{Get Link Columns
functions to determine if columns contain what appear to be links or emails}
\usage{
is_link_col(col)
}
\description{
Get Link Columns
functions to determine if columns contain what appear to be links or emails
}
\keyword{internal}
