% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pilotdata.R
\docType{data}
\name{pilotdata}
\alias{pilotdata}
\title{Development dataset}
\format{A dataframe.
\describe{
  \item{variables}{Todo}
  }}
\usage{
pilotdata
}
\description{
Description goes here..
}
\keyword{datasets}
