% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pilotdata.R
\docType{data}
\name{pilotdata}
\alias{pilotdata}
\title{Development dataset}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 812 rows and 33 columns.}
\usage{
pilotdata
}
\description{
Description goes here..
}
\keyword{datasets}
