% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sys_map.R
\name{sys_map}
\alias{sys_map}
\title{Map the studies}
\usage{
sys_map(studies_data, latitude, longitude)
}
\arguments{
\item{studies_data}{A dataframe of studies.}

\item{latitude}{Specify which column contains the latitude in the `studies_data` dataframe.}

\item{longitude}{Specify which column contains the latitude in the `studies_data` dataframe.}
}
\description{
Create a geographical map of the studies using `leaflet`.
}
