% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sys_map_shapefile.R
\name{sys_map_shapefile}
\alias{sys_map_shapefile}
\title{Plot systematic map from shapefile inputs}
\usage{
sys_map_shapefile(shp, popups = "")
}
\arguments{
\item{shp}{shapefile input in to eviatlas}

\item{popups}{Column with popup information}
}
\description{
Created For ES Hackathon 2018
}
\keyword{SystematicReview}
